refactor: /my-page layout 파일 정리
