+++ src/app/my-page/config/page.tsx
@@ -1,31 +1,22 @@
 import 'server-only';
 
-import { getCommunityList } from '~/api/domain/community.api';
+import { getCommunityListServer } from '~/api/domain/community.api.server';
+import { getUserInfoServer } from '~/api/domain/user.api.server';
 import { Divider } from '~/components/Divider';
-import { TopNavigation } from '~/components/TopNavigation';
 import { PlanetMenu } from '~/modules/PlanetMenu';
 import { UserMenu } from '~/modules/UserMenu';
 
 export const dynamic = 'force-dynamic';
 
 const MyPageConfig = async () => {
-  // TODO: userId 수정 필요
-  const userId = 1;
+  const { userProfileDto } = await getUserInfoServer();
 
-  const { communityListDtos } = await getCommunityList(userId);
+  const { communityListDtos } = await getCommunityListServer(userProfileDto.userId);
 
   const isBelongToCommunity = communityListDtos.length !== 0;
 
   return (
     <main>
-      <TopNavigation>
-        <TopNavigation.Left>
-          <TopNavigation.BackButton />
-        </TopNavigation.Left>
-        <TopNavigation.Title>
-          <h1 className="text-h5 font-semibold text-black">설정</h1>
-        </TopNavigation.Title>
-      </TopNavigation>
       <div className="pt-28pxr">
         {isBelongToCommunity && (
           <>
